Get rid of boxing inside Async<>
We want Async<T> to always impl Unpin. Turns out, boxing the inner value
is unnecessary for this; and as we never pin the inner value, we can just
implement Unpin for ourselves explicitly.

See https://doc.rust-lang.org/std/pin/index.html#pinning-is-not-structural-for-field
on why this is safe and sound :)