impl IoSafe for std::io::PipeReader & std::io::PipeWriter (#237)
It would be handy to have async wrappers for the pipe types (from
std::io::pipe()), so impl IoSafe for both.

The pipe types were introduced in 1.87, but the MSRV for async-io
appears to be 1.63 currently. So, protect the impls with cfg()s
and use autocfg to detect the type availability, similar to what was done
in a20076f ("Implement I/O-safe traits (#84)")

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>